Allow player to manage system daemons.

+++ Source/ProjectOglowia/Private/SystemControlCommand.cpp
@@ -35,6 +35,7 @@
 #include "PeacenetWorldStateActor.h"
 #include "SystemUpgrade.h"
 #include "CommonUtils.h"
+#include "DaemonManager.h"
 
 // Command name: systemctl
 // Description: Allows the player to poweroff or reboot the system, as well as manage system daemons.
@@ -50,6 +51,16 @@
 //  systemctl stop <service>  - stops the specified system daemon.
 void ASystemControlCommand::NativeRunCommand(UConsoleContext* InConsole, TArray<FString> InArguments)
 {
+    UDaemonManager* DaemonManager = nullptr;
+
+    if(!this->GetUserContext()->GetDaemonManager(DaemonManager))
+    {
+        InConsole->SetForegroundColor(EConsoleColor::Red);
+        InConsole->WriteLine(NSLOCTEXT("SystemCtl", "NeedsRoot", "systemctl: need to be root."));
+        InConsole->ResetFormatting();
+        this->Complete();
+        return;
+    }
     if(this->ArgumentMap["poweroff"]->AsBoolean())
     {
         InConsole->WriteLine(NSLOCTEXT("System", "Goodbye", "Goodbye."));
@@ -61,6 +72,46 @@ void ASystemControlCommand::NativeRunCommand(UConsoleContext* InConsole, TArray<
     {
         this->GetUserContext()->GetPeacenet()->QuitGame();
         return;
+    }
+    else if(this->ArgumentMap["list-units"]->AsBoolean())
+    {
+
+    }
+    else if(this->ArgumentMap["enable"]->AsBoolean())
+    {
+        FName DaemonName = FName(*this->ArgumentMap["<service>"]->AsString());
+
+        if(this->GetUserContext()->GetComputer().DisabledDaemons.Contains(DaemonName))
+        {
+            this->GetUserContext()->GetComputer().DisabledDaemons.Remove(DaemonName);
+        }
+    }
+    else if(this->ArgumentMap["disable"]->AsBoolean())
+    {
+        FName DaemonName = FName(*this->ArgumentMap["<service>"]->AsString());
+
+        if(!this->GetUserContext()->GetComputer().DisabledDaemons.Contains(DaemonName))
+        {
+            this->GetUserContext()->GetComputer().DisabledDaemons.Add(DaemonName);
+        }
+    }
+    else if(this->ArgumentMap["start"]->AsBoolean())
+    {
+        FName DaemonName = FName(*this->ArgumentMap["<service>"]->AsString());
+
+        DaemonManager->StartDaemonByName(DaemonName);
+    }
+    else if(this->ArgumentMap["stop"]->AsBoolean())
+    {
+        FName DaemonName = FName(*this->ArgumentMap["<service>"]->AsString());
+
+        DaemonManager->StopDaemonByName(DaemonName);
+    }
+    else if(this->ArgumentMap["restart"]->AsBoolean())
+    {
+        FName DaemonName = FName(*this->ArgumentMap["<service>"]->AsString());
+
+        DaemonManager->RestartDaemonByName(DaemonName);
     }
     this->Complete();
 }
